More refactoring
 * Separate all specific CSV ugliness into its own class
 * Support workspace:layer format
 * Dependency injection to choose WFS interface
   * Currently only ugly http implemented
