+++ src/extension/ncwms/src/main/java/au/org/emii/geoserver/wms/Ncwms.java
@@ -1,9 +1,7 @@
 package au.org.emii.geoserver.wms;
 
-import au.com.bytecode.opencsv.CSVReader;
 import org.apache.commons.io.IOUtils;
 import org.geotools.util.logging.Logging;
-import org.joda.time.DateTime;
 import org.json.JSONObject;
 
 import java.io.*;
@@ -27,53 +25,39 @@ import javax.xml.parsers.ParserConfigurationException;
 public class Ncwms {
     static Logger LOGGER = Logging.getLogger("au.org.emii.geoserver.wms.ncwms");
 
-    private static String timeFieldName = "time";
-    private static String urlFieldName= "file_url";
-
     public static Map<String, String> urlSubstitutions = new HashMap<String, String>();
-    private static String wfsServer = "http://localhost:8080/geoserver/ows";
 
-    public Ncwms() {}
-
-    public void setUrlFieldName(String urlFieldName) { Ncwms.urlFieldName = urlFieldName; }
-    public String getUrlFieldName() { return urlFieldName; }
-
-    public void setTimeFieldName(String timeFieldName) { Ncwms.timeFieldName = timeFieldName; }
-    public String getTimeFieldName() { return timeFieldName; }
-
-    public void setWfsServer(String wfsServer) { Ncwms.wfsServer = wfsServer; }
-    public String getWfsServer() { return wfsServer; }
+    private final URLIndexInterface urlIndexInterface;
 
     public void setUrlSubstitutions(Map<String, String> urlSubstitutions) { Ncwms.urlSubstitutions = urlSubstitutions; }
     public Map<String, String> getUrlSubstitutions() { return urlSubstitutions; }
 
+    public Ncwms(URLIndexInterface urlIndexInterface) {
+        this.urlIndexInterface = urlIndexInterface;
+    }
+
     public void getMetadata(HttpServletRequest request, HttpServletResponse response)
             throws ServletException, IOException {
         LOGGER.log(Level.INFO, "GetMetadata");
-        final String layerAndVariable = request.getParameter("layerName"); // TOOD needs to be case insensitive
-        String variable = layerAndVariable.split("/")[1];
-        final String wfsLayer = layerAndVariable.split("/")[0];
+        final LayerDescriptor layerDescriptor = new LayerDescriptor(request.getParameter("layerName")); // TOOD needs to be case insensitive
         final String item = request.getParameter("item");
 
         if (item != null && item.compareTo("timesteps") == 0) {
             String day = request.getParameter("day");
-            String cqlFilter = cqlFilterForSameDay(day, getTimeFieldName());
-            String urlParameters = getWfsUrlParameters(wfsLayer, timeFieldName, cqlFilter);
 
             JSONObject resultJson = new JSONObject();
-            resultJson.put("timesteps", getTimesForDay(wfsQuery(urlParameters)));
+            resultJson.put("timesteps", urlIndexInterface.getTimesForDay(layerDescriptor, day));
             response.getOutputStream().write(resultJson.toString().getBytes());
         }
         else if (item != null && item.compareTo("layerDetails") == 0) {
             LOGGER.log(Level.INFO, "Returning all available dates");
-            String urlParameters = getWfsUrlParameters(wfsLayer, timeFieldName, null);
 
-            Document getCapabilitiesDocument = getCapabilitiesXml(wfsLayer);
+            Document getCapabilitiesDocument = getCapabilitiesXml(layerDescriptor);
 
             JSONObject resultJson = new JSONObject();
-            resultJson.put("datesWithData", getUniqueDates(wfsQuery(urlParameters)));
-            resultJson.put("supportedStyles", getSupportedStyles(getCapabilitiesDocument, variable));
-            resultJson.put("palettes", getPalettes(getCapabilitiesDocument, variable));
+            resultJson.put("datesWithData", urlIndexInterface.getUniqueDates(layerDescriptor));
+            resultJson.put("supportedStyles", getSupportedStyles(getCapabilitiesDocument, layerDescriptor.variable));
+            resultJson.put("palettes", getPalettes(getCapabilitiesDocument, layerDescriptor.variable));
             response.getOutputStream().write(resultJson.toString().getBytes());
         }
     }
@@ -88,20 +72,18 @@ public class Ncwms {
         proxyWmsRequest(request, response);
     }
 
-    public void proxyWmsRequest(HttpServletRequest request, HttpServletResponse response)
+    private void proxyWmsRequest(HttpServletRequest request, HttpServletResponse response)
             throws ServletException, IOException {
-        String layerAndVariable = request.getParameter("LAYERS"); // TOOD needs to be case insensitive
-        String wfsLayer = layerAndVariable.split("/")[0];
-        String variable = layerAndVariable.split("/")[1];
+        LayerDescriptor layerDescriptor = new LayerDescriptor(request.getParameter("LAYERS")); // TOOD needs to be case insensitive
 
         String time = request.getParameter("TIME");
 
-        String wmsUrlStr = getWmsUrl(wfsLayer, time);
+        String wmsUrlStr = getWmsUrl(layerDescriptor, time);
 
         try {
             Map<String, String[]> wmsParameters = new HashMap(request.getParameterMap());
             wmsParameters.remove("TIME");
-            wmsParameters.put("LAYERS", new String[] { variable });
+            wmsParameters.put("LAYERS", new String[] { layerDescriptor.variable });
 
             String queryString = encodeMapForRequest(wmsParameters);
 
@@ -163,11 +145,11 @@ public class Ncwms {
         }};
     }
 
-    private Document getCapabilitiesXml(String wfsLayer)
+    private Document getCapabilitiesXml(LayerDescriptor layerDescriptor)
         throws IOException {
         String wmsUrl = null;
         try {
-            wmsUrl = getWmsUrl(wfsLayer, null);
+            wmsUrl = getWmsUrl(layerDescriptor, null);
             String getCapabilitiesUrl =
                 String.format("%s?service=WMS&version=1.3.0&request=GetCapabilities",
                     wmsUrl
@@ -189,127 +171,8 @@ public class Ncwms {
         }
     }
 
-    private String cqlFilterForTimestamp(String timestamp, String timeFieldName) {
-        LOGGER.log(Level.INFO, String.format("Returning cql for timestamp '%s'", timestamp));
-        return String.format("%s = %s", timeFieldName, timestamp);
-    }
-
-    private String cqlFilterForSameDay(String day, String timeFieldName) {
-        String timeStart = day;
-        String timeEnd = getNextDay(timeStart);
-        LOGGER.log(Level.INFO, String.format("Returning times of day '%s'", day));
-
-        return String.format(
-            "%s >= %s AND %s < %s",
-            timeFieldName, timeStart,
-            timeFieldName, timeEnd
-        );
-    }
-
-    private String getWfsUrlParameters(String wfsLayer, String propertyName, String cqlFilter) throws UnsupportedEncodingException {
-        String urlParameters =
-            String.format(
-                "typeName=%s&SERVICE=WFS&outputFormat=csv&REQUEST=GetFeature&VERSION=1.0.0&PROPERTYNAME=%s",
-                wfsLayer, propertyName
-            );
-
-        if (cqlFilter != null && ! cqlFilter.isEmpty()) {
-            urlParameters += String.format("&CQL_FILTER=%s", URLEncoder.encode(cqlFilter, StandardCharsets.UTF_8.name()));
-        }
-
-        return urlParameters;
-    }
-
-    private String getTimeFromDate(String date) {
-        DateTime jodaDate = new DateTime(date);
-        return jodaDate.toLocalTime().toString() + "Z";
-    }
-
-    private String getNextDay(String date) {
-        DateTime jodaDate = new DateTime(date);
-        return jodaDate.plusDays(1).toString();
-    }
-
-    public CSVReader processCsvInput(InputStream csv) throws IOException {
-        CSVReader csvReader = new CSVReader(new InputStreamReader(csv, StandardCharsets.UTF_8.name()));
-
-        csvReader.readNext(); // Skip first line, it's the header
-        return csvReader;
-    }
-
-    public Map<Integer, Map<Integer, Set<Integer>> > getUniqueDates(InputStream csv) throws IOException {
-        CSVReader csvReader = processCsvInput(csv);
-
-        Map<Integer, Map<Integer, Set<Integer> > > dates =
-            new HashMap<Integer , Map<Integer, Set<Integer> > >();
-
-        for (String[] currentRow = csvReader.readNext();
-            currentRow != null;
-            currentRow = csvReader.readNext()) {
-
-            LOGGER.log(Level.INFO, String.format("Processing row '%s'", currentRow));
-
-            DateTime date = new DateTime(currentRow[1]);
-
-            Integer year = date.getYear();
-            Integer month = (date.getMonthOfYear() - 1); // Zero indexed
-            Integer day = date.getDayOfMonth();
-
-            if (!dates.containsKey(year)) {
-                dates.put(year, new HashMap<Integer, Set<Integer> >());
-            }
-
-            if (!dates.get(year).containsKey(month)) {
-                dates.get(year).put(month, new HashSet<Integer>());
-            }
-
-            dates.get(year).get(month).add(day);
-        }
-
-        return dates;
-    }
-
-    public List<String> getTimesForDay(InputStream csv) throws IOException {
-        CSVReader csvReader = processCsvInput(csv);
-
-        List<String> timesOfDay = new ArrayList<String>();
-
-        for (String[] currentRow = csvReader.readNext();
-            currentRow != null;
-            currentRow = csvReader.readNext()) {
-
-            LOGGER.log(Level.INFO, String.format("Processing row '%s'", currentRow));
-            String time = getTimeFromDate(currentRow[1]);
-            timesOfDay.add(time);
-        }
-
-        return timesOfDay;
-    }
-
-    public InputStream wfsQuery(String urlParameters) throws IOException {
-        URL url = new URL(wfsServer + "?" + urlParameters);
-        LOGGER.log(Level.INFO, String.format("WFS query '%s'", url));
-        return url.openConnection().getInputStream();
-    }
-
-    public String getWmsUrl(String wfsLayer, String time) throws IOException {
-        // By default form a query to get the last file ordered by timestamp (reverse)
-        String extraUrlParameters = "&" + String.format("sortBy=%s+D", timeFieldName); // Sort by time, descending
-        String cqlFilter = null;
-        if (time != null && time != "") {
-            cqlFilter = cqlFilterForTimestamp(time, timeFieldName);
-            extraUrlParameters = "";
-        }
-
-        String urlParameters = getWfsUrlParameters(wfsLayer, timeFieldName, cqlFilter);
-        urlParameters += "&" + "maxFeatures=1";
-        urlParameters += extraUrlParameters;
-
-        CSVReader csvReader = processCsvInput(wfsQuery(urlParameters));
-
-        String wmsUrl = "";
-        String[] currentRow = csvReader.readNext();
-        return mangleUrl(currentRow[1]);
+    private String getWmsUrl(LayerDescriptor layerDescriptor, String time) throws IOException {
+        return mangleUrl(urlIndexInterface.getUrlForTimestamp(layerDescriptor, time));
     }
 
     private String mangleUrl(String url) {
